refactor(chat): extract rememberChatSearchState shared by both screens
The desktop and mobile group screens each carried a near-identical ~100-line
block of search state, derived values, and the matcher / prefetch / "search
older" dig / watchdog effects plus the search-bar slot. Move it into one
rememberChatSearchState holder (uiComposeMain) returning the bits the screen
needs (active, currentHitId, scrollNonce, hitIds, onToggle, bar). Removes ~265
lines of duplication and a class of desktop/mobile drift.
